requirements.txt
discord.py~=2.5.2
python-dotenv~=1.0.1
neo4j~=5.28.1
psycopg[binary,pool]
table2ascii

